Refactor : 보너스 숫자 예외 처리 메소드를 BonusNumberException.class로 변경
