+++ src/main/java/lotto/exception/BonusNumberException.java
@@ -1,0 +1,30 @@
+package lotto.exception;
+
+import lotto.model.BonusNumber;
+import lotto.model.Number;
+
+import java.util.List;
+
+import static lotto.view.ErrorMessage.duplicatedBonusNumber;
+
+public class BonusNumberException extends ExceptionController{
+
+    public static BonusNumber checkBonusNumberException(List<Integer> numbers, int bonus) throws IllegalArgumentException {
+        try {
+            checkNumberException(bonus);
+            checkBonusDuplicateException(numbers, bonus);
+            return new BonusNumber(bonus);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public static void checkBonusDuplicateException(List<Integer> numbers, int bonus) throws IllegalArgumentException {
+        try {
+            Number.checkDuplicate(numbers, bonus);
+        } catch (IllegalArgumentException e) {
+            duplicatedBonusNumber();
+            throw new IllegalArgumentException();
+        }
+    }
+}
